Add token auth create, delete, search memo

+++ src/main/java/com/mapping/mapping/memo/memoController.java
@@ -1,7 +1,10 @@
 package com.mapping.mapping.memo;
 
+import com.mapping.mapping.user.security.TokenProvider;
+
 import java.util.List;
 import java.util.Optional;
+import java.util.Collections;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -21,88 +24,119 @@ import java.io.File;
 import java.io.IOException;
 
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.servlet.http.HttpServletRequest;
 
 // CROS 설정 - 개빡치는놈
 @CrossOrigin("*")
 @RestController
-@RequestMapping(value = "/memo")
-public class memoController {
+@RequestMapping(value = "/api/memo")
+public class memoController{
 	private memoRepository boardRep;
-
-	@Value("${C_SERVER_IMGSAVE_ADDR}")
-	private String SERVER_IMGSAVE_ADDR;
-
-	@Value("${C_SERVER_DOMAIN}")
-	private String SERVER_DOMAIN;
 	
 	@Autowired	
 	public memoController(memoRepository boardRep) {
 		this.boardRep = boardRep;
 	}
 
-    // URL POST 메모 생성 : POST(enctype="multipart/form-data") -> {url}/upload
+	@Value("${C_SERVER_IMGSAVE_ADDR}")
+    private String SERVER_IMGSAVE_ADDR;
+
+    @Value("${C_SERVER_DOMAIN}")
+    private String SERVER_DOMAIN;
+
+	@Autowired
+    private TokenProvider tokenProvider;
+
+	@PostMapping("/token")
+	public String tokentest(HttpServletRequest request) {
+		try {
+			String token = parseBearerToken(request);
+			if (token != null && !token.equalsIgnoreCase("null")) {
+				String userNickname = tokenProvider.validate(token);
+	
+				return userNickname;
+			}
+		} catch (Exception exception) {
+			exception.printStackTrace();
+			return "Exception occurred: " + exception.getMessage();
+		}
+		return "An error occurred";
+	}
+
+ 	// URL POST 메모 생성 : POST(enctype="multipart/form-data") -> {url}/upload
     @PostMapping("/upload")
     public String createMemo(
-        @RequestParam("file") MultipartFile file,
+        HttpServletRequest request,
+        @RequestParam(value = "file",required = false) MultipartFile file,
         @RequestParam("content") String content,
-        @RequestParam("writer") String writer,
         @RequestParam("lat") String lat,
         @RequestParam("lng") String lng,
         @RequestParam("tag") String tag) throws IOException {
+            
+            try{
+                String token = parseBearerToken(request);
+                if(token != null && !token.equalsIgnoreCase("null")){
+                    String userNickname = tokenProvider.validate(token);
 
-			LocalDateTime nowDate = LocalDateTime.now();
-	
-			String nowDate_DB = nowDate.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss"));
-			String nowDate_Img = nowDate.format(DateTimeFormatter.ofPattern("yyyy-MM-dd_HH:mm:ss"));
-			String nowDate_fd = nowDate.format(DateTimeFormatter.ofPattern("yyyy-MM-dd"));
+            if (userNickname == null){
+                //돌아가
+            }else{
+            LocalDateTime nowDate = LocalDateTime.now();
+    
+            String nowDate_DB = nowDate.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss"));
+            String nowDate_Img = nowDate.format(DateTimeFormatter.ofPattern("yyyy-MM-dd_HH:mm:ss"));
+            String nowDate_fd = nowDate.format(DateTimeFormatter.ofPattern("yyyy-MM-dd"));
 
-			//이미지 업로드 경로 지정(절대경로)
+            //이미지 업로드 경로 지정(절대경로)
             String uploadPath = SERVER_IMGSAVE_ADDR + "/"+ nowDate_fd + "/";
 
-			File UploadFolder = new File(uploadPath);
+            File UploadFolder = new File(uploadPath);
 
-			 //만약 해당 날짜 폴더가 없다면 생성
-			if (!UploadFolder.exists()) {
-				try{
-		    		UploadFolder.mkdir();
-	        		} 
-	        		catch(Exception e){
-		    		e.getStackTrace();
-				}        
-         		}else {
-			}
+             //만약 해당 날짜 폴더가 없다면 생성
+            if (!UploadFolder.exists()) {
+                try{
+                    UploadFolder.mkdir();
+                    } 
+                    catch(Exception e){
+                    e.getStackTrace();
+                }        
+                }else {
+            }
 
-			String filename = file.getOriginalFilename();
-
-			memo item = new memo();
+            memo item = new memo();
             item.setContent(content);
-            item.setWriter(writer);
+            item.setWriter(userNickname);
             item.setLat(lat);
             item.setLng(lng);
             item.setDate(nowDate_DB);
             item.setTag(tag);
 
-			// 만약 이미지파일이 들어오지않았을때 null 처리
-			if (filename == ""){
-				String img = "null";
-				item.setImg(img);
+            // 만약 이미지파일이 들어오지않았을때 null 처리
+            if (file == null || file.isEmpty()){
+                String img = "null";
+                item.setImg(img);
 
-			}else{
-				File dest = new File(uploadPath + nowDate_Img + "_" + writer + ".jpg");
-            	file.transferTo(dest);
+            }else{
+                File dest = new File(uploadPath + nowDate_Img + "_" + userNickname + ".jpg");
+                file.transferTo(dest);
 
-				//저장된 이미지 불러오는 서버 주소
-            	String img = "http://" + SERVER_DOMAIN + ":8080/images/" + nowDate_fd + "/" + nowDate_Img + "_" + writer + ".jpg";
-           		item.setImg(img);
-			}
-			
+                //저장된 이미지 불러오는 서버 주소
+                String img = "http://" + SERVER_DOMAIN + ":8080/images/" + nowDate_fd + "/" + nowDate_Img + "_" + userNickname + ".jpg";
+                item.setImg(img);
+            }
+            
             boardRep.save(item);
 
-			
-			System.out.println("\n" + nowDate_Img + ") 메모 생성 완료");
+            
+            System.out.println("\n" + nowDate_Img + ") 메모 생성 완료");
             return "INFO) Memo create Success";
-        }
-
+                }
+            }
+            }catch(Exception exception){
+                exception.printStackTrace();
+            }
+        return "INFO) Memo create fail"; // 예외 발생 시 반환 값
+    }   
 	
     //메모 전체 로드 : {url}/memo - 작동
 	@GetMapping
@@ -138,17 +172,35 @@ public class memoController {
     	return boardRep.findByContentContains(content);
 	}
 
-	//작성자검색 : {url}/memo/search?writer={writer}
+	//자기가 만든 메모만 불러오기(With token) : {url}/memo/search
 	@GetMapping(value = "/search")
-	public List<memo> searchByWriter(@RequestParam String writer) {
+	public List<memo> searchByWriter(HttpServletRequest request) {
+ 	   try {
+  	    	String token = parseBearerToken(request);
+   	    	if (token != null && !token.equalsIgnoreCase("null")) {
+        	    String userNickname = tokenProvider.validate(token);
+        	    if (userNickname != null) {
+            	    LocalDateTime nowDate = LocalDateTime.now();
+            	    String nowDate_Img = nowDate.format(DateTimeFormatter.ofPattern("yyyy-MM-dd_HH:mm:ss"));
 
-		LocalDateTime nowDate = LocalDateTime.now();
-		String nowDate_Img = nowDate.format(DateTimeFormatter.ofPattern("yyyy-MM-dd_HH:mm:ss"));
+					System.out.println("\n" + nowDate_Img + ") 작성자 기준 메모 검색 완료");
+            	    return boardRep.findByWriter(userNickname);
+            	}
+        	}
+    	} catch (Exception exception) {
+        	exception.printStackTrace();
+        	// 예외 발생 시 메시지를 반환
+        	//List<memo> errorResponse = new ArrayList<>();
+        	//memo errorMessageMemo = new memo();
+        	//errorMessageMemo.setContent("Exception occurred: " + exception.getMessage());
+        	//errorResponse.add(errorMessageMemo);
+        	//return errorResponse;
+    	}
 
-		System.out.println("\n" + nowDate_Img + ") 작성자 기준 메모 검색 완료");
-
-    	return boardRep.findByWriter(writer);
+    	// 여기까지 도달하는 경우, 예외 또는 조건에 부합하지 않음을 나타냄
+    	return Collections.emptyList(); // 빈 리스트 반환
 	}
+
 
 	//태그검색 : {url}/memo/tagsearch?tag={tag}
 	@GetMapping(value = "/tagsearch")
@@ -188,13 +240,48 @@ public class memoController {
 
     //메모삭제 : {url}/memo/delete?id={id}
 	@GetMapping(value = "/delete")
-	public void deleteById(@RequestParam(value = "id") Long id) {
-    	boardRep.deleteById(id);
-
-		LocalDateTime nowDate = LocalDateTime.now();
-		String nowDate_Img = nowDate.format(DateTimeFormatter.ofPattern("yyyy-MM-dd_HH:mm:ss"));
-
-		System.out.println("\n" + nowDate_Img + ") ID : " + id + " 메모 삭제 완료");
+	public String deleteById(HttpServletRequest request, @RequestParam(value = "id") Long id) {
+		try {
+			String token = parseBearerToken(request);
+			if (token != null && !token.equalsIgnoreCase("null")) {
+				String userNickname = tokenProvider.validate(token);
+	
+				LocalDateTime nowDate = LocalDateTime.now();
+				String nowDate_Img = nowDate.format(DateTimeFormatter.ofPattern("yyyy-MM-dd_HH:mm:ss"));
+	
+				System.out.println("\n" + nowDate_Img + ") ID : " + id + " 메모 삭제 시도");
+	
+				Optional<memo> memoOptional = boardRep.findById(id);
+				if (memoOptional.isPresent()) {
+					String writer = memoOptional.get().getWriter();
+	
+					if (writer.equals(userNickname)) {
+						boardRep.deleteById(id);
+						System.out.println("메모 삭제 완료");
+						return "메모 삭제 완료";
+					} else {
+						System.out.println("작성자와 로그인 사용자가 다릅니다.");
+						return "작성자와 로그인 사용자가 다릅니다.";
+					}
+				} else {
+					System.out.println("해당 ID의 메모가 없습니다.");
+					return "해당 ID의 메모가 없습니다.";
+				}
+			}
+		} catch (Exception exception) {
+			exception.printStackTrace();
+		}
+		return "An error occurred";
 	}
+	
+    //Request Header에서 Authorization 필드의 Bearer Token을 가져오는 메서드
+    private String parseBearerToken(HttpServletRequest request){
+        String bearerToken = request.getHeader("Authorization");
+
+        //문자를 가지고 있는지 bearerToken이 맞는지 확인, 맞으면 7번째 문자부터 가져온다.
+        if(org.springframework.util.StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer "))
+            return bearerToken.substring(7);
+        return null;
+    }
 
 }
